third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-csp-expected.txt

CONSOLE ERROR: Refused to load the image 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE' because it violates the following Content Security Policy directive: "img-src http://127.0.0.1:8000".

CONSOLE ERROR: Refused to load the image 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE' because it violates the following Content Security Policy directive: "img-src http://127.0.0.1:8000".

This is a testharness.js-based test.
PASS Verify CSP control of fetch() in a Service Worker
Harness: the test ran to completion.
